Testing debug

+++ src/trainer/weak-classifier.cpp
@@ -9,7 +9,7 @@ namespace violajones
   bool WeakClassifier::check(Window win, std::shared_ptr<IntegralImage> image)
   {
     auto feature_value = feature_->compute_value(win.top_left, win.size_ratio, image);
-    int sized_value = feature_value / (win.size_ratio * win.size_ratio);
+    int sized_value = (int) (feature_value / (win.size_ratio * win.size_ratio));
     auto normalized_value = Feature::normalize_feature(sized_value, win.deviation);
     return parity_ * normalized_value < parity_ * threshold_;
   }
